A tad bit of clean up and an entry for future tests
Need to make a website config file. One feature it will
have for sure is to set files that should be not be loaded
as part of the WebFS and maybe set the root page. Should
also add a server config. One feature will be to set the
websites root folder.